gather common functions
