+++ yamanaka/nextjs-app-dir/app/articles/new/page.tsx
@@ -11,6 +11,7 @@ import {
   Button,
   FormErrorMessage,
 } from "@/app/_common/components";
+import { getCookie, getCsrfToken } from "@/app/_common/fucntions";
 
 export default function CreateArticle() {
   const router = useRouter();
@@ -47,30 +48,6 @@ export default function CreateArticle() {
       }
     }
   };
-
-  async function getCsrfToken() {
-    const res = await fetch("http://localhost:8000/sanctum/csrf-cookie", {
-      cache: "no-store",
-      credentials: "include",
-    });
-    return res;
-  }
-  async function getCookie(key: string) {
-    const cookies = document.cookie.split("; ");
-    const check = async (key: string, str: string) => {
-      const arr = str.split("=");
-      const res = key == arr[0] ? decodeURIComponent(arr[1]) : false;
-      return res;
-    };
-    let ans = "";
-    const res = await Promise.all(
-      cookies.map(async (cookie) => {
-        const found = await check(key, cookie);
-        if (found != false) ans = found;
-      })
-    );
-    return ans;
-  }
 
   return (
     <div>
